Move quorum sig verification to crypto
